Move classes from core to business package
