+++ src/main/java/business/Album.java
@@ -1,4 +1,4 @@
-package core;
+package business;
 
 public class Album{
