+++ src/main/java/business/AlbumManagerSingleton.java
@@ -1,4 +1,4 @@
-package core;
+package business;
 
 import java.lang.reflect.Constructor;
 import java.lang.reflect.InvocationTargetException;
